Assignment-1
Assignment-1: Pass-1 of a Two-Pass Assembler

This assignment implements Pass-1 of a two-pass assembler for a hypothetical machine. The program generates a Symbol Table and Intermediate Code file while processing sample instructions and assembler directives. The assembler also handles forward references and error-checking during the assembly process.

Question : 
Design suitable data structures and implement Pass-1 of a two-pass assembler for a hypothetical machine.
Generate a symbol table and an intermediate code file. The implementation should consider the following:

1.Sample instructions from each category and a few assembler directives.
2.Forward references.
3.Error handling: symbols used but not defined, invalid instructions/registers, etc.